--- src/Cli.Tests/CliTests.cs
@@ -1,37 +1,0 @@
-using System.CommandLine;
-using System.CommandLine.Parsing;
-using Drift.Cli.Abstractions;
-using Drift.Cli.Commands.Scan;
-using Microsoft.Extensions.Logging.Abstractions;
-
-namespace Drift.Cli.Tests;
-
-public class CliTests {
-  //TODO fix
-  [Explicit]
-  [Test]
-  public async Task HostTest() {
-    // Arrange
-    var rootCommand = new RootCommand();
-    rootCommand.Subcommands.Add( new ScanCommand( new NullLoggerFactory() ) );
-    //TODO rootCommand.AddScanCommand();
-
-    var console = new TestConsole();
-
-    // Act
-    var result = await CommandLineParser.Parse(
-      rootCommand,
-      "scan",
-      new CommandLineConfiguration( rootCommand ) { Output = console.Out, Error = console.Error }
-    ).InvokeAsync();
-
-    // Assert
-    var outputText = console.Out.ToString().Trim();
-    using ( Assert.EnterMultipleScope() ) {
-      Assert.That( outputText, Is.Not.Empty );
-      Assert.That( result, Is.EqualTo( ExitCodes.Success ) );
-    }
-
-    await Verify( outputText );
-  }
-}
