+++ src/Cli.Tests/Commands/InitCommandTests.GeneratedSpecWithDiscoveryIsValid.verified.txt
