+++ src/Cli.Tests/Commands/InitCommandTests.GeneratedSpecWithoutDiscoveryIsValid.verified.txt
