+++ src/Cli.Tests/Commands/InitCommandTests.MissingNameOption.verified.txt
