+++ src/Cli.Tests/Commands/LintCommandTests.LintInvalidSpec_specName=network_single_device_host_outputFormat=normal.verified.txt
