+++ src/Cli.Tests/Commands/LintCommandTests.LintValidSpec_specName=network_single_subnet_outputFormat=.verified.txt
