+++ src/Cli.Tests/Commands/ScanCommandTests.NonExistingSpecOption.verified.txt
